0006/requirements.txt
streamlit
openai
pydantic_ai
python-dotenv
nest_asyncio
rich
